feat: 🎸 changed folder structure to match create-t3-app
✅ Closes: BUC-10
