+++ package.json
@@ -21,7 +21,6 @@
     "@trpc/next": "^10.9.0",
     "@trpc/react-query": "^10.9.0",
     "@trpc/server": "^10.9.0",
-    "@types/prettier": "^2.7.2",
     "daisyui": "^2.49.0",
     "next": "13.1.6",
     "next-auth": "^4.18.10",
@@ -32,6 +31,7 @@
   },
   "devDependencies": {
     "@types/node": "^18.11.18",
+    "@types/prettier": "^2.7.2",
     "@types/react": "^18.0.27",
     "@types/react-dom": "^18.0.10",
     "@typescript-eslint/eslint-plugin": "^5.49.0",
