+++ src/pages/_app.tsx
@@ -2,7 +2,7 @@ import { type AppType } from 'next/app';
 import { type Session } from 'next-auth';
 import { SessionProvider } from 'next-auth/react';
 
-import { trpc } from '../utils/trpc';
+import { api } from '../utils/api';
 
 import '../styles/globals.css';
 import NiceModal from '@ebay/nice-modal-react';
@@ -20,4 +20,4 @@ const MyApp: AppType<{ session: Session | null }> = ({
   );
 };
 
-export default trpc.withTRPC(MyApp);
+export default api.withTRPC(MyApp);
