+++ src/pages/api/auth/[...nextauth].ts
@@ -4,7 +4,7 @@ import DiscordProvider from 'next-auth/providers/discord';
 import { PrismaAdapter } from '@next-auth/prisma-adapter';
 
 import { env } from '../../../env/server.mjs';
-import { prisma } from '../../../server/db/client';
+import { prisma } from '../../../server/db';
 
 export const authOptions: NextAuthOptions = {
   // Include user.id on session
@@ -23,7 +23,15 @@ export const authOptions: NextAuthOptions = {
       clientId: env.DISCORD_CLIENT_ID,
       clientSecret: env.DISCORD_CLIENT_SECRET,
     }),
-    // ...add more providers here
+    /**
+     * ...add more providers here
+     *
+     * Most other providers require a bit more work than the Discord provider.
+     * For example, the GitHub provider requires you to add the
+     * `refresh_token_expires_in` field to the Account model. Refer to the
+     * NextAuth.js docs for the provider you want to use. Example:
+     * @see https://next-auth.js.org/providers/github
+     */
   ],
 };
