+++ src/pages/api/restricted.ts
@@ -1,6 +1,6 @@
-import { type NextApiRequest, type NextApiResponse } from "next";
+import { type NextApiRequest, type NextApiResponse } from 'next';
 
-import { getServerAuthSession } from "../../server/common/get-server-auth-session";
+import { getServerAuthSession } from '../../server/auth';
 
 const restricted = async (req: NextApiRequest, res: NextApiResponse) => {
   const session = await getServerAuthSession({ req, res });
@@ -8,12 +8,12 @@ const restricted = async (req: NextApiRequest, res: NextApiResponse) => {
   if (session) {
     res.send({
       content:
-        "This is protected content. You can access this content because you are signed in.",
+        'This is protected content. You can access this content because you are signed in.',
     });
   } else {
     res.send({
       error:
-        "You must be signed in to view the protected content on this page.",
+        'You must be signed in to view the protected content on this page.',
     });
   }
 };
