+++ src/pages/api/trpc/[trpc].ts
@@ -1,17 +1,19 @@
-import { createNextApiHandler } from "@trpc/server/adapters/next";
+import { createNextApiHandler } from '@trpc/server/adapters/next';
 
-import { env } from "../../../env/server.mjs";
-import { createContext } from "../../../server/trpc/context";
-import { appRouter } from "../../../server/trpc/router/_app";
+import { env } from '../../../env/server.mjs';
+import { createTRPCContext } from '../../../server/api/trpc';
+import { appRouter } from '../../../server/api/root';
 
 // export API handler
 export default createNextApiHandler({
   router: appRouter,
-  createContext,
+  createContext: createTRPCContext,
   onError:
-    env.NODE_ENV === "development"
+    env.NODE_ENV === 'development'
       ? ({ path, error }) => {
-          console.error(`❌ tRPC failed on ${path}: ${error}`);
+          console.error(
+            `❌ tRPC failed on ${path ?? '<no-path>'}: ${error.message}`,
+          );
         }
       : undefined,
 });
