+++ src/server/api/routers/example.ts
@@ -1,13 +1,12 @@
-import { z } from "zod";
+import { z } from 'zod';
+import { createTRPCRouter, publicProcedure } from '../trpc';
 
-import { router, publicProcedure } from "../trpc";
-
-export const exampleRouter = router({
+export const exampleRouter = createTRPCRouter({
   hello: publicProcedure
     .input(z.object({ text: z.string().nullish() }).nullish())
     .query(({ input }) => {
       return {
-        greeting: `Hello ${input?.text ?? "world"}`,
+        greeting: `Hello ${input?.text ?? 'world'}`,
       };
     }),
   getAll: publicProcedure.query(({ ctx }) => {
