--- src/server/trpc/router/auth.ts
@@ -1,7 +1,0 @@
-import { publicProcedure, router } from '../trpc';
-
-export const authRouter = router({
-  getSession: publicProcedure.query(({ ctx }) => {
-    return ctx.session;
-  }),
-});
